Update carousel for general nubmer of picrtures
Except the carousel for the home page, that have specific photos with specific descriptions for each  ìphoto (detailed accessibility for blind people), all the other carousels (for performers and artistic events) are now able to display any number of pictures.
The only side effect is that, assuming photos are named as 1.jpg, 2.jpg, ..., n.jpg, the loop stops when tries to access the picture named n+1.jpg but get 404 message instead.
But the 404 message is not visible to the user thus does not affect the user experience in any way, unless they are navigating the page with the Chrome DevTools open (or developer tools of other browsers).