Update signupFrontend.js
If there is a problem in the automatic sign in after the registration, the user is not shown the message of "something went wrong" anymore, that made the user think that the registration was not successfull, even if it was. Instead, now the user receives the message that confirm their registration and then is sent to the signin page.